CampusKartBackend-master/requirements.txt
cmake==3.22.2
distlib==0.3.1
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
firebase-admin==5.2.0
flake8==3.9.2
future==0.18.2
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
httplib2==0.20.4
identify==2.3.0
instaloader==4.8.2
jedi==0.17.0
joblib==1.0.1
json5==0.9.5
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.3
names==0.3.0
nodeenv==1.6.0
numpy==1.20.2
olefile==0.46
pandas==1.3.5
pathspec==0.7.0
phonenumbers==8.12.43
Pillow==9.1.0
pipreq==0.4
pipreqs==0.4.11
pkginfo==1.7.0
pre-commit==2.15.0
proto-plus==1.20.3
protobuf==3.19.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycosat==0.6.3
pyflakes==2.3.1
python-dotenv==0.16.0
PyYAML==5.4.1
pyzmq==20.0.0
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
QtPy==1.9.0
rsa==4.8
scikit-learn==0.24.1
scipy==1.6.2
sqlparse==0.4.2
uritemplate==4.1.1
virtualenv==20.4.3
webencodings==0.5.1
whitenoise==6.0.0
SQLAlchemy==1.4.27